Split test file into multiple by feature
