+++ test/replace-term_test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var grunt = require('grunt');
+
+module.exports = {
+
+};
